backend/requirements.txt
flask
opencv-python
numpy
torch
facenet-pytorch
onnxruntime
Pillow
